Perubahan Awan Bensin DLL

+++ FileUnity Baong/Project Kurir/Assets/Scriptable Object For Saving/SaveUIValue.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 [CreateAssetMenu]
 public class SaveUIValue : ScriptableObject
 {
-    public int JumlahUang;
+    public float JumlahUang;
 
     public int MisiYangAktif;
